refactor: make appointment
