+++ src/components/SectionSchedule/SectionSchedule.tsx
@@ -1,26 +1,14 @@
 import { Box, Typography, Chip, Accordion, AccordionSummary, AccordionDetails } from "@mui/material";
 import { IAppointments } from "../../api/schedule";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import { Appointment } from "../../api/patients";
-import { Modal } from "../Modal/Modal";
-import { useState } from "react";
-import { MakeAppointmentModal } from "./MakeAppointmentModal";
 
-export const SectionSchedule = ({ data, onClick, doctor }: { data: IAppointments; onClick: (day: string, time: string) => Promise<Appointment | undefined | void> | void; doctor: string }) => {
+interface IProps {
+  data: IAppointments;
+  onClick: () => void;
+  onClickAppointment?: (date: string, time: string) => void;
+}
 
-  const [isOpen, setOpen] = useState(false);
-  const [day, setDay] = useState("");
-  const [time, setTime] = useState("");
-  const [doctorName, setDoctorName] = useState("");
-
-  const openModal = (date: string, time: string) => {
-    setOpen(true);
-    setDay(date);
-    setTime(time);
-    setDoctorName(doctor);
-  };
-  const closeModal = () => setOpen(false);
-
+export const SectionSchedule = ({ data, onClick, onClickAppointment }: IProps) => {
 
   return (
     <Box>
@@ -41,28 +29,14 @@ export const SectionSchedule = ({ data, onClick, doctor }: { data: IAppointments
                 )
                 .map(([time, patientId]) =>
                   patientId !== null ? (
-                    <Chip key={`${date}-${time}`} sx={{ cursor: "pointer" }} label={time} onClick={() => {
-                      openModal(date, time);
-                      // onClick(date, time);
-                    }} />
+                    <Chip key={`${date}-${time}`} sx={{ cursor: "pointer" }} label={time} onClick={() => onClickAppointment ? onClickAppointment(date, time) : onClick()} />
                   ) : (
-                    <Chip key={`${date}-${time}`} sx={{ cursor: "pointer", margin: "5px" }} label={time} variant="outlined" onClick={() => {
-                      openModal(date, time);
-                      // onClick(date, time);
-                    }} />
+                    <Chip key={`${date}-${time}`} sx={{ cursor: "pointer", margin: "5px" }} label={time} variant="outlined" onClick={() => onClickAppointment ? onClickAppointment(date, time) : onClick()} />
                   ),
                 )}
             </AccordionDetails>
           </Accordion>
         ))}
-      <Modal
-        open={isOpen}
-        onClose={closeModal}
-        aria-labelledby="modal-modal-title"
-        aria-describedby="modal-modal-description"
-      >
-        <MakeAppointmentModal actionFunc={onClick} date={day} time={time} doctor={doctorName} close={closeModal} />
-      </Modal>
     </Box>
   );
 };
